Keep renderer and writer in single file
